First pass at store interface
Has full FileStore/MemoryStore implementation for Messages. For
file stores, the command line "-filestore" points to the root
directory where all the channels will be stored. Messages are
persisted and can be recovered after a restart.
The interface for Subscription is there with callouts, but this
part is work-in-progress. Things could change.