Update test to demonstrate the issue
A test already existed but assumed that the current behavior was
correct. Adding a check (currently commented out) to ensure that
no unwanted redelivery occurs after a server restart.

Refactor tests to use helper functions to check on various things.

Changed the test default datastore for each package (server and
stores) to use a temporary directory.

Resolves #56