Fix DATA RACE when getting the subStore from a ChannelStore
We set the subStore (a server's specific structure) in a UserData
of a store's ChannelStore. This was done in two steps, which was
opening a door to data races: a go routine could lookup a channel
that was just created in another but for which the UserData field
was not yet set.
Replaced store's LookupOrCreateChannel() to CreateChannel() which
accepts the user data as parameter and fails if channel already
exists. This allows the server code to turn around and return
the result of LookupChannel().

Resolves #84