Add auto-compaction of clients file
The client file is a serie of "add" or "delete" or client records
(records are two strings, the client ID and heart beat inbox).
In order to reduce the size of the file, when the number of
active clients is below a certain threshold, the file is recreated
with only "add" records of the currently active clients.