[CHANGED] Redelivery is now always forced
When a consumer was stalled (number of ack pending is >= of MaxInflight)
when a redelivery occurred, the server would be prevented from sending
messages since the number of outstanding acks was already at the limit.
Some code was introduced to force redelivery after a certain number
of failed attempts. This code is now removed and redelivery always
resend messages for which acks have not been received.

This helps with message ordering on server or durable restart.

Resolves #187