[FIXED] Durable Queue Subscriber not removed from store when leaving the group
Which then caused after a server restart to re-incarnate as a "ghost"
durable queue subscriber.
This PR makes sure that the member is removed from store as it should,
but also attempts to not recover such ghost queue durable on restart.
This does not however "cleanup" the store file, just prevents the
runtime state to be incorrect.

Resolves #215