[UPDATED] Redelivery constraints for Queue Group
If a member of a queue group crashes, the server will not detect
that the client is gone until after the number of failed heartbeats.
In the meantime, messages may be assigned to that member. After
the AckWait period elapses, the server attempts to redeliver the
unacknowledged messages for that member. If there is at least
a failed heartbeat, the server would skip redelivery. This is
fine for regular subscribers, but in the case of a queue group,
it is likely that the message is actually redelivered to a different
member. So this PR skips the heartbeat check for a subscriber
that is part of a queue group, unless the group is down to that
single member.

Addresses #220