[FIXED] Message not sent and stalled subscriber if MaxInflight <= 0
The server was not checking MaxInflight values from the SubscriptionRequest.
If value was 0 (or less), even if a message was available, the
server would try to send but because the number of pendind Acks (0)
would be >= of the MaxInflight (<=0), then no message would be sent
and subscriber would be flagged as stalled.
The server will now return an error if MaxInflight is <= 0.

Resolves #226