[FIXED] Ensure message timestamp increases monotonically
Since we use `time.Now()` and this is returning the "wall clock",
it is subject to jump both forward or backwards, either due to
user changing the time, NTP, or leap second.
There is no much we can do, but at least we ensure that the messages
in a given log have their timestamp increasing monotonically.
That is, if M1 is stored before M2, then:
   M1.Sequence<M2.Sequence && M1.Timestamp <= M2.Timestamp.

The expiration code also now accounts for possible negative delta
between "now" and a given message timestamp.

In part addresses: #231