Reduce memory usage for FileMsgStore
By not keeping msgIndex (which contains message's offset on
data file, timestamp and message size), the memory footprint
is now reduced and actually does not increase for a given
message store, regardless of the number of messages stored.
This causes however to access the index file - to retrieve
the message offset - in order to lookup a message, during
message expiration, or enforcing limits.
The FileMsgStore still uses a small message cache which
contains messages that have been just written or looked-up.