[CHANGED] RunServer() and RunServerWithOpts() APIs to return error
Those APIs used to only return an instance of StanServer. If there
was any issue on startup, the call would panic. The purpose of this
change is to instead return an error. The executable will also
display the error as a fatal error and exit (without the panic).

This would require a change in the Go client tests and for users
running the streaming server programmatically.