Added BackoffTimeCheck utility (#269)
* Added BackoffTimeCheck utility

In situations where an error/trace is printed but there is a
possibility that the print is too frequent, which could cause the
log to quickly fill up, this utility should be used.

You first create a BackoffTimeCheck object giving a minimum and maximum
frequency and a factor at which the frequency is increased.

An Ok() function tells you if it is ok to issue the print statement.
When returning true, the object sets its next print target which
causes subsequent calls to Ok() to return false if they are made
too soon.

The frequency allowed to print is increased by the given factor
anytime the function returns Ok(). It is capped by the max frequency
set when creating the object.

After reaching the max frequency, if a call to Ok() is made after
more than 3x the max frequency since the last call that returned
true, then the call returns Ok() but is also auto-reset.

There is also a Reset() API to explicitly reset the object
to an initial state.