Do not panic when opening a file removed from file manager
This can happen when the store is being closed. Msgs and Subs stores
are closed first and then the file manager is closed. When stores
are closed, files are removed from the file manager (but manager
status is still opened). So when trying to open a file that is
no longer in the file manager map, return an error instead of panic.