Do not panic when opening a file removed from file manager
The fix issue on #273 was not right since it was possible for the
state to be changed to fileRemoved but the file still in the map
when another go routine would have the file manager lock in
openFile. The check to see if file was present in the map would
return true, but the state would be fileRemoved, which would still
cause the panic.