[FIXED] Use of single config file with TLS section
All but the "tls" options are not overlapping with NATS, so using
the content of a single configuration file with Streaming or NATS
would work well.
However, "tls" option exists on both and NATS Server rejects
unknown fields.
A way to address this (and selected in this PR) is to add a top-level
streaming map to the NATS Streaming server configuration. It makes
it then possible to mix NATS and Streaming configuration in the
same configuration file and have "tls" sections not conflicting.
The top-level "streaming" map is not required, unless you need
to have NATS "tls" and Streaming "tls" present in the same file.

Resolves #333