Allow AckWait to be specified in ms for tests and some other tweaks
- When a subscription is created, it can specify an AckWait() which
is a duration, but the library sends the value as a number of
seconds. For tests, this can lengthen the time the test suite takes.
Making a change that allows to specify a negative value that means
that server interprets as milliseconds. The global boolean is
used to trigger this behavior.
- Changed tests to use smaller AckWait().
- Some FT tests sleep times have been reduced.
- Create explicit NATS connection with nats.ReconnectWait() for
  some tests where server is restarted. It shorten the test time
  because the client can reconnect faster.