[FIXED] Override of NATS related command line parameters
The override using natsd.MergeOptions() does not work very well in
that it is not possible to override options if the command line
parameter is actually a default scalar value (false for boolean,
0 for int, etc...). Using the reflection as we did for stan's
options to solve that.

Also, fix #336 introduced a bug that caused some of the NATS command
lines parameters to be ignored (if they were not explicitly covered
in MergeOptions). This issue is not in any public release, but was
in master since #336 was merged.

Again, using overrideWithCmdLineParams() here solve both problems.