[IMPROVED] Partitioning behavior on topology change
When a server running in partitioning mode detects that a server
joined the NATS cluster, it checks again that its channels do not
conflict with channels of other servers in this cluster. To
give a chance to the newly joined NATS Server in the cluster to
propagate its subscriptions interest, a delay is added before
sending the list and checking for replies.