Fixed panic when running tests on Windows network drive
The raft log was located in current directory, which when running
on a network drive on Windows was crashing without any indication
that this was the caused.
Changing call to ioutil.TempDir() to use "" instead of ".", which
on Windows will then create the files in the Windows temp directory
which by default is in C:\. This solves the panic.