Update/Add test to check for client connection
When a client exits without calling Connection.Close(), the server
does not know and if the client is restarted with same client ID,
the server will attempt to communicate with the last known HBInbox
of the registered client. If the client fails to respond, it will
close the known client connection and replace with the new client
info.

Updated the existing test and added one for Clustering that shows
there is currently a bug that does not properly handle this
situation.