[ADDED] Encryption at rest for messages
Added a new store that adds encryption of messages payload capabilities
to existing store implementations.
Server will use encryption for any selected store type if started
with the Encrypt option set to true and the EncryptionKey string.
The environment variable NATS_STREAMING_ENCRYPTION_KEY will be used
if set. If not, will use the EncryptionKey option.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>