[ADDED] Automatic deletion of channels after a certain inactivity
Added MaxInactivity as a new channel limit. By default, this limit
is set to unlimited. If the user sets this limit, then when
a channel does not have any active subscription and no new message
for at least MaxInactivity duration, the server will delete such
channel.

Note that offline (queue) durable subscriptions do not prevent
deletion of a channel.

Resolves #358