[FIXED] Clustering: possible panic on startup
This could happen when applying raft log entries to the state machine
when s.raft was not yet created. The fix could have been to move
s.raf.leader to the server structure. Instead, I have refactored
the creation of the raft node a bit so we can use the leader and
initialized atomic from s.raft instead of StanServer.
Also made s.raft.shutdown() idempotent.