[FIXED] Restore from snapshot
When the raft node is created, local snapshots if present will be
first restored. We need to detect the restore from init as opposed
to a restore due to a follower falling behind and catching up with
the leader. The previous use of an atomic was not correct since
there could be some delay that would make a restore from leader
think that the init was happening in NewRaft(). Fixing that with
knowledge of list of local snapshots and knowing that RAFT will
restore in NewRaft() the most recent local snapshot.

Also, a synchronization bug between leadershipAcquired() and the
ioLoop has been fixed. The defect was introduced in a recent commit
and not part of any release.

Resolves #503