[ADDED] Support for client pings
Server now listens to client PINGs and send a response back to
indicate if the client connection is still valid. Indeed, it
is possible for a client to be "disconnected" from the server
and another client (with same client ID) to replace it.
When the old client communication resumes, there would be two
clients with the same client ID. Once the old client sends a
PING, the server will respond that the connection is no longer
valid.
Also, published message now contain a connection ID that allows
the server to reject messages from the old client in the situation
described above.

This PR contains changes on the server. Clients implementations
to follow.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>