[UPDATED] Replaced github.com/boltdb with go.etcd.io/boltdb
The original boltdb repo has been archived for a while, so
moving to the currently maintained one.

Removed code that would re-create raft log, which in some case
was worse than the issue that we tried to originally fixed.
Using the newer boltdb options to not sync free list on disk
and use a map type instead of array.

Resolves #718

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>