[FIXED] Clustering: unable to complete leadrship actions
In some cases, it was possible for a follower to become leader
but be unable to complete all the actions to fully become leader.
The underlying issue was that if there were snapshots to send
to the followers, those were unable to restore the snapshot
because the new leader was not listening to those requests yet.
Also, there could be a risk that a channel is part of the
snapshot sent to followers but then would be deleted while
the followers would try to install the snapshot.

Resolves #762

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>