[FIXED] Panic on raft write issue
An user reported the panic due to a volume full issue and was
asking if we could print a more explicit error message.

It seems that there is an issue with Boltdb and a recent param
called NoFreelistSync. When set to true and a transaction is
rolled-back, a panic occurs.

This boolean was used when moving from boltdb to etcd/bbolt
in the recent 0.12.0 release.

Disabling for now until the underlying issue can be resolved.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>