[FIXED] Cluster: maintain sub's ID on replay
When a server in the cluster restarts and replays the subscription
create event, a new subscription ID was assigned to it. It is not
a problem per-se, but could make debugging using logs a bit more
challenging. Ensure that we keep the same subscription ID for
a given subscription regardless of cluster restart/snapshot.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>