[FIXED] Clustering: erroneous "too many subscriptions per channel"
This was introduced by #802. The server was replacing the sub's ID
after the call to CreateSub(), but that meant that DeleteSub() in
the underlying sub store would not find that subscription, which
will then cause it to report too many subscriptions after a while
because the removed subscriptions were not removed from their
internal map.
The approach in this PR is to make RaftSubStore completely bypass
the underlying SubStore. It should work with a mix of servers
of different versions.

Resolves #809

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>