[FIXED] Clustering: message store not flushed on followers
With the SQLStore, this would cause messages to be accumulated
in write cache causing memory growth and no message stored in the
DB until the follower shutsdown or become leader.

Resolves #846

Also made server tests skip SQL specific tests if running tests
with `-sql=false` for users that don't have/want to setup an
SQL server.

Resolves #845

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>