[FIXED] Raft logging
When RAFT logging is enabled, starting 0.15.0, it looks like
the raft logging is a bit off. Something similar to this:
```
[32605] 2019/07/25 13:46:57.787222 [WRN] STREAM:  raft: Heartbeat timeout from "" reached, starting election
[32605] 2019/07/25 13:46:57.787270 [INF] STREAM:  raft: Node at test-cluster.a.test-cluster [Candidate] entering Candidate state in term 2
[32605] 2019/07/25 13:46:59.792967 [WRN] STREAM:  raft: Election timeout reached, restarting election
[32605] 2019/07/25 13:46:59.793034 [INF] STREAM:  raft: Node at test-cluster.a.test-cluster [Candidate] entering Candidate state in term 3
[32605] 2019/07/25 13:46:59.793060 [ERR] STREAM: ] raft: Failed to make RequestVote RPC to {Voter b test-cluster.b.test-cluster}: nats: timeout
```
Notice the extra white space between `STREAM:` and `raft:` and
a weird `] raft:` trace. This is becaues 0.15.0 updated the RAFT
dependency and between v1.0.0 and current release, they changed
the way they do logging. So had to adapt our code that intercept
the RAFT logger to pipe it to the STAN logger. A test was added
to catch such change in the future.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>