[ADDED] In/Out Msgs/Bytes in /serverz monitoring endpoint
Unlike total_msgs/bytes that represent the total number of messages
stored, in_msgs/bytes and out_msgs/bytes are the messages received
by the server and sent to clients.
The `in_msgs` counter is updated when processing inbound client
messages.
The `out_msgs` counter is updated when sending a message to a subscription,
either as a new or redelivered message.

Note that the size for in_bytes/out_bytes is different because these
values represent the bytes received (a PubMsg protobuf) and sent
(a MsgProto protobuf) that have different overhead.

Resolves #974

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>