[FIXED] FileStore: Empty() should remove all message files (dat/idx)
Doing some manual tests with removing the single .dat file for
a channel, I realized that in cluster mode, when the channel message
store was being emptied, the .idx file was not removed, which then
caused a problem since the info was added to the index file while
it should have been replaced.
Ensure that all files matching the messages files pattern are removed
during an Empty() call.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>