[FIXED] Cluster: add flag to allow node to proceed on restore failure
When a node restore from a snapshot, it could be that the snapshot
information shows that a channel should have at least say messages
1 to 100. If the streaming store does not have these sequences,
the node is supposed to reconcile with the leader. However, if
all nodes are restarted and all have this issue, the cluster
would not be able to be restarted at all.
The new `cluster_proceed_on_restore_failure` will allow node(s)
to start.

Relates to #1010

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>