[ADDED] Clustering: ability to add/remove nodes at runtime
If option `allow_add_remove_node` in `streaming{cluster{}}` block
is enabled (or command line `--cluster_allow_add_remove_node`)
then the leader will listen to `_STAN.raft.<cluster id>.node.remove`
(and `...node.add`) to which an user can send a simple NATS request
with the value being the name of the node to add/remove.

If successful, the leader will return `+OK`, if not `-ERR` with
the error string explaining why it failed.

Note that removing a non-existent node will still return `+OK`.

If the leader receives a request to remove itself, it will work
but the process will then exit. It can be added back by sending
the request to the `..node.add` subject.

Note that this updates the RAFT configuration in the RAFT log but
does NOT change the configuration file. The admin will be responsible
to change the configuration on file so that when nodes are restarted
they reflect the new configuration (list of peers).

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>