[ADDED] Authentication fields in streaming configuration block
Abiltity to provide Username, Password or Token in the streaming
configuration block.

If the configuration is:
```
authorization {
  users [
    { user: foo, password: bar}
  ]
}
streaming {
  ..
}
```
the server would fail to start unless user passes `-user foo -pass bar`
from the command line. But there was no way to pass those from the
configuration file. You can now configure this way:

```
streaming {
  ..
  username: "foo"
  password: "bar"
}
```
in order for the server to be able to start without the additional
`-user` and `-pass` command line parameters.

Also adding the `token` configuration option.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>