[ADDED] NKey support
You can now configure the streaming server by providing a file
to the NKey seed. This will be used when creating the NATS connections
and when NKey authentication is required.
The file is read to get the seed in order to sign the nonce from
the NATS server. The seed is wiped out from memory after this
is done.

Resolves #1095

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>