[ADDED] Configuration parameters for RAFT's BoltDB store
This replaces PRs #1155 and #1156.

In 1155, I made boltdb's default freelist type to be map instead
of array, and in 1156 added options to enable array type.

This PR keeps the default to array and allow user to change to
map instead (so the config param is now `bolt_free_list_map` instead
of `bolt_free_list_array` that was introduced in 1156.

The reason is to minimize possible unexpected behavior in deployments
that did not have any issue.
I am not sure of all the extent (performance wise) of changing to
map instead of array and would rather it to be an explicit choice
made by users, and only if they observe a performance degradation
that can be attributed to the freelist.

We can always make it the default later.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>