[ADDED] FileStore: RecordSizeLimit to limit single record read size
In case of memory corruption, it is possible that the record size
is way greater than it should, which would cause the server to
create a buffer of the wrong size in the attempt to read the
record. This new option will limit how big the buffer needed
to read the record from disk can be.

Resolves #1255

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>