[CHANGED] Reject channels with different case (Foo vs foo)
Due to the way channels are stored to persistent storage, the
first spelling of a channel caused that channel to be persisted
under that name.
If an application later publishes/subscribes to the same channel
but with a different case (say first "Foo" then "foo"), the
server would treat it as a new channel but would still store
data in the first "Foo" channel, causing message sequences to
be reset and inability to consume after a server restart.

From now on, the server will reject published messages or new
subscriptions if it detects that a similar channel already exists
with a different "case".

Resolves #1265

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>