Bump to v0.25.2
A change in test framework in between releases has changed the
SQL setting for tests, which now was enabled by default even
for server tests. This caused the release process to fail.
I then pushed the 0.25.1 tag but forgot to first update the
version in server.go, which was correctly caught by the
test TestVersionMatchesTag which is used to validate and trigger
the release.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>